Refactored around literal query
