+++ trueseeing/core/android/db.py
@@ -7,7 +7,7 @@ from trueseeing.core.model.issue import Issue
 from trueseeing.core.tools import noneif
 
 if TYPE_CHECKING:
-  from typing import Any, Iterable, Tuple, Dict, Optional, Iterator, List
+  from typing import Any, Iterable, Tuple, Dict, Optional, Iterator, List, Set
   from trueseeing.core.android.store import Store
   from trueseeing.core.android.model.code import InvocationPattern
   from trueseeing.core.model.issue import IssueConfidence
@@ -309,6 +309,44 @@ class Query:
   def issues_by_group(self, *, detector: str, summary: str) -> Iterable[Issue]:
     for m in self.db.execute('select detector, summary, synopsis, description, seealso, solution, info1, info2, info3, confidence, cvss3_score, cvss3_vector, source, row, col from analysis_issues where detector=:detector and summary=:summary order by cvss3_score desc, confidence desc', dict(detector=detector, summary=summary)):
       yield self._issue_from_row(m)
+
+  def op_get(self, k: int) -> Optional[Op]:
+    for t,v in self.db.execute('select t,v from ops where op=?', (k, )):
+      return Op(t, v)
+    return None
+
+  def op_store_ops(self, ops: Iterable[Op], c: Any = None) -> None:
+    if c is None: c = self.db
+    c.executemany('insert into ops(op,idx,t,v) values (?,?,?,?)', ((t._id, t._idx, t.t, t.v) for t in ops))
+
+  def op_count_ops(self, c: Any = None) -> int:
+    if c is None: c = self.db
+    for cnt, in c.execute('select count(1) from ops where idx=0'):
+      return cnt # type: ignore[no-any-return]
+    return 0
+
+  def op_store_classmap(self, classmap: Set[Tuple[int, int]], c: Any = None) -> int:
+    if c is None: c = self.db
+    c.executemany('insert into ops_class(class,op) select ?,op from ops where op between ? and ?', ((start, start, end) for start, end in classmap))
+    return len(classmap)
+
+  def op_generate_methodmap(self, c: Any = None) -> int:
+    if c is None: c = self.db
+    detected_methods = 0
+    c.execute("create table tmp1 (op int primary key)")
+    c.execute("create table tmp2 (op int primary key)")
+    c.execute("insert into tmp1 select op from ops where t='directive' and v='method'")
+    c.execute("insert into tmp2 select a.op as op from ops as a left join ops as c on (a.op=c.op-c.idx) where a.t='directive' and a.v='end' and c.idx=1 and c.v='method'")
+    c.execute('insert into ops_method(method,op) select mm.sop as method,ops.op from (select tmp1.op as sop,(select min(op) from tmp2 where op>tmp1.op) as eop from tmp1) as mm left join ops on (ops.op between mm.sop and mm.eop)')
+    for cnt, in c.execute('select count(1) from tmp1'):
+      detected_methods = cnt
+    c.execute("drop table tmp1")
+    c.execute("drop table tmp2")
+    return detected_methods
+
+  def op_finalize(self) -> None:
+    from trueseeing.core.android.db import StorePrep
+    StorePrep(self.db).stage2()
 
   @classmethod
   def _issue_confidence_to_int(cls, c: IssueConfidence) -> int:
